+++ src/components/MediaList.vue
@@ -1,53 +1,54 @@
 <template>
   <div>
-    <!-- <waterfall :line-gap="width" :min-line-gap="width" :single-max-width="width" :watch="items" align="center">
-      <waterfall-slot
-        v-for="item in items"
-        :width="width"
-        :height="300"
-        :key="item.media._id"
-      > -->
-    <q-btn label="Pack" color="primary" @click.native="pack" />
-    <bricks
-      ref="bricks"
-      :data="items"
-      :sizes="sizes"
-    >
-      <template slot-scope="scope">
-        <q-card color="primary" inline v-bind:class="[scope.item.type] + 'l'" class="media-cardl" @click="openItem(scope.item.media._id)">
-          <q-card-media v-if="scope.item.type == 'book' || scope.item.type == 'movie' || scope.item.type == 'video' || scope.item.type == 'article' || scope.item.type == 'image'">
-            <img :src="scope.item.media.thumbURL" />
-          </q-card-media>
-          <q-card-title v-if="scope.item.type == 'book' || scope.item.type == 'movie' || scope.item.type == 'video' || scope.item.type == 'article' || scope.item.type == 'discourse' || scope.item.type == 'note'">
-              {{ scope.item.media.title }}
-              <span v-if="scope.item.type == 'book' || scope.item.type == 'movie' || scope.item.type == 'video' || scope.item.type == 'article' || scope.item.type == 'discourse'" v-for="author in scope.item.media.author" :key="author" slot="subtitle">{{ author }}</span>
-              <span v-if="scope.item.type == 'note'" slot="subtitle">{{ scope.item.media.text }}</span>
-            </q-card-title>
-            <q-card-main v-if="scope.item.type == 'quote'">
-              <p>{{ scope.item.media.text }}</p>
-              <p class="q-body-2"><span v-for="author in scope.item.media.mediaid.author" :key="author">{{ author }} | </span>{{ scope.item.media.mediaid.title }}</p>
-            </q-card-main>
-        </q-card>
-      </template>
-    </bricks>
-      <!-- </waterfall-slot>
-    </waterfall> -->
-    <q-modal v-model="resourceOpen">
-      <resource-preview :type="resourceType" :id="resourceID" />
+    <div class="row gutter-sm">
+      <div class="col-12">
+        <q-btn label="Pack" color="primary" @click.native="pack" />
+      </div>
+      <div class="col-12">
+        <q-select
+          multiple
+          v-model="selectedTypes"
+          :options="types"
+          dark
+        />
+      </div>
+      <div class="col-12">
+        <bricks
+          ref="bricks"
+          :data="showItems"
+          :sizes="sizes"
+        >
+          <template slot-scope="scope">
+            <q-card color="primary" inline v-bind:class="[scope.item.type] + 'l'" class="media-cardl" @click.native="openItem(scope.item.media, scope.item.type)">
+              <q-card-media v-if="scope.item.type == 'book' || scope.item.type == 'movie' || scope.item.type == 'video' || scope.item.type == 'article' || scope.item.type == 'image'">
+                <img :src="scope.item.media.thumbURL" />
+              </q-card-media>
+              <q-card-title v-if="scope.item.type == 'book' || scope.item.type == 'movie' || scope.item.type == 'video' || scope.item.type == 'article' || scope.item.type == 'discourse' || scope.item.type == 'note'">
+                  {{ scope.item.media.title }}
+                  <span v-if="scope.item.type == 'book' || scope.item.type == 'movie' || scope.item.type == 'video' || scope.item.type == 'article' || scope.item.type == 'discourse'" v-for="author in scope.item.media.author" :key="author" slot="subtitle">{{ author }}</span>
+                  <span v-if="scope.item.type == 'note'" slot="subtitle">{{ scope.item.media.text }}</span>
+                </q-card-title>
+                <q-card-main v-if="scope.item.type == 'quote'">
+                  <p>{{ scope.item.media.text }}</p>
+                  <p class="q-body-2"><span v-for="author in scope.item.media.mediaid.author" :key="author">{{ author }} | </span>{{ scope.item.media.mediaid.title }}</p>
+                </q-card-main>
+            </q-card>
+          </template>
+        </bricks>
+      </div>
+    </div>
+    <q-modal v-model="resourceOpen" content-classes="resource-modal">
+      <resource-preview :type="resourceType" :resource="resource" />
     </q-modal>
   </div>
 </template>
 
 <script>
-// import Waterfall from 'vue-waterfall/lib/waterfall'
-// import WaterfallSlot from 'vue-waterfall/lib/waterfall-slot'
 import Bricks from 'vue-bricks'
 import ResourcePreview from 'components/ResourcePreview.vue'
 
 export default {
   components: {
-    // Waterfall,
-    // WaterfallSlot,
     Bricks,
     ResourcePreview
   },
@@ -55,10 +56,11 @@ export default {
   props: ['items', 'width'],
   data () {
     return {
+      showItems: [],
       isImage: this.type === 'images',
       resourceOpen: false,
       resourceType: '',
-      resourceID: '',
+      resource: '',
       sizes: [
         { columns: 1, gutter: 20 },
         { mq: '800px', columns: 2, gutter: 20 },
@@ -67,29 +69,79 @@ export default {
       ],
       cardStyle: {
         width: this.width
-      }
+      },
+      selectedTypes: [
+        'book'
+      ],
+      types: [
+        {
+          label: 'Book',
+          value: 'book'
+        },
+        {
+          label: 'Movie',
+          value: 'movie'
+        },
+        {
+          label: 'Image',
+          value: 'image'
+        },
+        {
+          label: 'Video',
+          value: 'video'
+        },
+        {
+          label: 'Article',
+          value: 'article'
+        },
+        {
+          label: 'Note',
+          value: 'note'
+        },
+        {
+          label: 'Quote',
+          value: 'quote'
+        },
+        {
+          label: 'Outline',
+          value: 'outline'
+        },
+        {
+          label: 'Idea',
+          value: 'idea'
+        }
+      ]
     }
   },
   watch: {
-    'items': function () {
+    'selectedTypes': function () {
+      this.showItems = this.items.filter(this.checkType)
+    },
+    'items': function (value) {
+      this.showItems = value.filter(this.checkType)
       this.pack()
     },
     'width': function () {
       this.pack()
     }
   },
+  // mounted () {
+  //   this.showItems = this.items.filter(this.checkType)
+  // },
   updated () {
     console.log('updated')
     console.log(this.$refs.bricks)
     this.pack()
   },
   methods: {
-    openItem (id, type) {
-      console.log(id)
+    openItem (item, type) {
+      console.log(item)
       console.log(type)
       switch (type) {
         case 'book':
-          // this.$router.push({ name: 'book', params: { id: id } })
+          this.resource = item
+          this.resourceType = type
+          this.resourceOpen = true
           break
         case 'movie':
           // this.$router.push({ name: 'movie', params: { id: id } })
@@ -116,6 +168,9 @@ export default {
     },
     pack () {
       this.$refs.bricks.pack()
+    },
+    checkType (item) {
+      return this.selectedTypes.includes(item.type)
     }
   }
 }
@@ -154,6 +209,19 @@ export default {
 @media screen and (min-width: 1500px) {
   .bookl, .moviel, .imagel, .videol, .articlel, .notel, .quotel {
     width: 250px;
+  }
+}
+
+.resource-modal {
+  /* padding: 30px; */
+  width: 100%;
+  height: 100%;
+}
+
+@media screen and (min-width: 1200px) {
+  .resource-modal {
+    min-width: 650px;
+    width: 650px;
   }
 }