+++ src/components/QuoteList.vue
@@ -1,22 +1,58 @@
 <template>
   <div>
     <!-- <q-btn color="negative" @click="fixQuotes">Fix Quotes</q-btn> -->
-    <h4>Snippet List <small><q-icon color="primary" name="fa-plus" @click.native="openAdd" class="cursor-pointer" /> <q-icon color="primary" name="fa-toggle-down" @click.native="toggleQuotes" class="cursor-pointer" /></small></h4>
-    <q-list separator multiline v-if="showQuotes">
-      <quote-list-item
-        v-for="quote in quotes"
-        v-bind:quoteObj="quote"
-        v-bind:key="quote.id"
-        location
-        tags
-        bible
-        notes
+    <h4>Snippet List <small><q-icon color="primary" name="fa-plus" @click.native="openAdd" class="cursor-pointer" /> <q-icon color="primary" name="fa-toggle-down" @click.native="toggleQuotes" class="cursor-pointer" /></small>
+      <span v-if="loading">
+        &nbsp;
+        <q-spinner color="primary" />
+      </span>
+    </h4>
+    <div v-if="showQuotes">
+      <h5>Ideas</h5>
+      <q-list separator multiline>
+        <idea-list-item
+          v-for="idea in ideas"
+          v-bind:ideaObj="idea"
+          v-bind:key="idea.id"
+          location
+          tags
+          bible
+          notes
+        />
+      </q-list>
+      <h5>Outlines</h5>
+      <bricks
+        ref="bricks"
+        :data="outlines"
+        :sizes="isModal ? subSizes : sizes"
       >
-      </quote-list-item>
-    </q-list>
+        <template slot-scope="scope">
+          <outline-list-item
+            v-bind:outlineObj="scope.item"
+            v-bind:key="scope.item.id"
+            location
+            tags
+            bible
+            notes
+          />
+        </template>
+      </bricks>
+      <h5>Quotes</h5>
+      <q-list separator multiline>
+        <quote-list-item
+          v-for="quote in quotes"
+          v-bind:quoteObj="quote"
+          v-bind:key="quote.id"
+          location
+          tags
+          bible
+          notes
+        />
+      </q-list>
+    </div>
     <q-modal v-model="addOpen" content-classes="add-quote-modal">
       <!-- <q-icon name="fa-close" size="2rem" @click.native="closeAdd" class="float-right cursor-pointer" /> -->
-      <q-tabs position="top" align="center" no-pane-border>
+      <q-tabs v-model="tabSelection" position="top" align="center" no-pane-border>
         <q-tab default slot="title" label="Quote" name="quote-tab" icon="fa-quote-right" />
         <q-tab slot="title" label="Outline" name="outline-tab" icon="fa-list-ul" />
         <q-tab slot="title" label="Idea" name="idea-tab" icon="fa-lightbulb" />
@@ -36,31 +72,56 @@
 </template>
 
 <script>
+import Bricks from 'vue-bricks'
 import QuoteListItem from 'components/QuoteListItem.vue'
+import OutlineListItem from 'components/OutlineListItem.vue'
+import IdeaListItem from 'components/IdeaListItem.vue'
 import QuoteForm from 'components/QuoteForm.vue'
 import OutlineForm from 'components/OutlineForm.vue'
 import IdeaForm from 'components/IdeaForm.vue'
 
 export default {
   components: {
+    Bricks,
     QuoteListItem,
+    OutlineListItem,
+    IdeaListItem,
     QuoteForm,
     OutlineForm,
     IdeaForm
   },
-  props: ['mediaid', 'media', 'mediaType'],
+  props: ['mediaid', 'media', 'mediaType', 'modal'],
   data () {
     return {
       id: this.mediaid,
       type: this.mediaType,
       mediaObj: this.media,
       quotes: [],
+      outlines: [],
+      ideas: [],
       addOpen: false,
       showQuotes: false,
-      isMobile: this.$q.platform.is.mobile
+      loading: false,
+      tabSelection: 'quote-tab',
+      isMobile: this.$q.platform.is.mobile,
+      isModal: (this.modal === ''),
+      sizes: [
+        { columns: 1, gutter: 20 },
+        { mq: '800px', columns: 2, gutter: 20 },
+        { mq: '1260px', columns: 3, gutter: 20 },
+        { mq: '1600px', columns: 4, gutter: 20 }
+      ],
+      subSizes: [
+        { columns: 1, gutter: 20 },
+        { mq: '1200px', columns: 2, gutter: 20 }
+      ]
     }
   },
   watch: {
+    mediaid (value) {
+      this.id = value
+      this.init()
+    },
     media (value) {
       this.mediaObj = value
     },
@@ -68,27 +129,62 @@ export default {
       this.type = value
     }
   },
-  mounted () {
-    this.init()
-    console.log(this.$q.platform)
-  },
+  // mounted () {
+  //   this.init()
+  // },
   methods: {
     init () {
-      this.database.snippets('quotes', this.id, (data) => {
-        this.quotes = data
+      this.showQuotes = false
+      this.loading = true
+      this.database.snippets('all', this.id, (data) => {
+        console.log(data)
+        this.quotes = data[0]
+        this.outlines = data[1]
+        this.ideas = data[2]
+        this.showQuotes = true
+        this.loading = false
+        // if (callback) callback()
       })
     },
     openAdd () {
       this.addOpen = true
-      this.$refs.quoteForm.init(true)
+      switch (this.tabSelection) {
+        case 'quote-tab':
+          this.$refs.quoteForm.init(true)
+          break
+        case 'outline-tab':
+          this.$refs.outlineForm.init(true)
+          break
+        case 'idea-tab':
+          this.$refs.ideaForm.init(true)
+          break
+        default:
+          console.log('wrong tab...')
+      }
     },
-    closeAdd () {
+    closeAdd (newItem, type) {
       this.addOpen = false
-      // Call init to reload quotes with any updates
-      this.init()
+      switch (type) {
+        case 'quote':
+          this.quotes.push(newItem)
+          break
+        case 'outline':
+          this.outlines.push(newItem)
+          break
+        case 'idea':
+          this.ideas.push(newItem)
+          break
+        default:
+          console.log('wrong new add...')
+      }
     },
     toggleQuotes () {
-      if (this.showQuotes) this.showQuotes = false; else this.showQuotes = true
+      if (this.showQuotes) {
+        this.showQuotes = false
+      } else {
+        this.loading = true
+        this.init()
+      }
     }
   }
 }